# optimze: reuse subset entry to optimze alloc/inuse memory